refactor(dify-ui): consolidate Vitest configuration
