+++ packages/dify-ui/.gitignore
@@ -1,3 +1,4 @@
 /coverage
+/.vitest-browser
 /dist
 /storybook-static
